Fix various issues from the backlog, incl. float parsing

+++ json-utils/src/lib.rs
@@ -231,14 +231,14 @@ pub struct TimeIntervalValue {
 
 #[derive(Serialize, Deserialize, Clone, PartialEq, Debug)]
 pub struct AmountOfMoneyValue {
-    pub value: f32,
+    pub value: f64,
     pub precision: Precision,
     pub unit: Option<String>,
 }
 
 #[derive(Serialize, Deserialize, Clone, PartialEq, Debug)]
 pub struct TemperatureValue {
-    pub value: f32,
+    pub value: f64,
     pub unit: Option<String>,
 }
